streamlit_app/requirements.txt
numpy==1.21.2
pandas==1.3.3
streamlit==0.85.0
Pillow==8.3.2
streamlit_drawable_canvas==0.8.0


